[Add]テスト対象オブジェクトを修正し、divideで３を2で割った結果を取得するテストを作成

+++ app/src/main/java/com/misakikawaguchi/tutorial/Tutorial.kt
@@ -8,7 +8,7 @@ class Calculator {
     }
 
     // 割り算
-    fun divide(x: Int, y: Int): Int {
-        return x / y
+    fun divide(x: Int, y: Int): Double {
+        return x / y.toDouble()
     }
 }
